+++ server/src/infrastructure/repositories/exchange-db.repository.ts
@@ -3,7 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { CreateExchangeDTO } from '../../infrastructure/controllers/exchange/exchange.create.dto';
 import { UpdateExchangeDTO } from '../../infrastructure/controllers/exchange/exchange.update.dto';
-import { IExchangeDbRepository } from '../../domain/repositories/exchange-db.repository.interface';
+import { IExchangeDbRepository } from '../../domain/repositories/exchange-db.repository';
 import { Exchange, ExchangeDocument } from '../entities/exchange.entity';
 
 @Injectable()
