+++ server/src/usecases/exchange/delete-exchange.usecase.ts
@@ -1,4 +1,4 @@
-import { IExchangeDbRepository } from '../../domain/repositories/exchange-db.repository.interface';
+import { IExchangeDbRepository } from '../../domain/repositories/exchange-db.repository';
 
 class DeleteExchangeUseCase {
   constructor(private readonly exchangeDbRepository: IExchangeDbRepository) {}
